Start to move more basic tasks into puffin-dispatch
Starting to come together

Now it compiles
