+++ crates/puffin-cli/src/commands/pip_sync.rs
@@ -210,8 +210,8 @@ pub(crate) async fn sync_requirements(
         downloads
             .into_iter()
             .partition_map(|download| match download {
-                puffin_installer::Download::Wheel(wheel) => Either::Left(wheel),
-                puffin_installer::Download::SourceDist(sdist) => Either::Right(sdist),
+                puffin_distribution::Download::Wheel(wheel) => Either::Left(wheel),
+                puffin_distribution::Download::SourceDist(sdist) => Either::Right(sdist),
             });
 
     // Build any missing source distributions.
