+++ crates/puffin-dispatch/Cargo.toml
@@ -11,13 +11,14 @@ authors = { workspace = true }
 license = { workspace = true }
 
 [dependencies]
+distribution-types = { path = "../distribution-types" }
 gourgeist = { path = "../gourgeist" }
 pep508_rs = { path = "../pep508-rs" }
 platform-host = { path = "../platform-host" }
 platform-tags = { path = "../platform-tags" }
 puffin-build = { path = "../puffin-build" }
 puffin-client = { path = "../puffin-client" }
-distribution-types = { path = "../distribution-types" }
+puffin-distribution = { path = "../puffin-distribution" }
 puffin-installer = { path = "../puffin-installer" }
 puffin-interpreter = { path = "../puffin-interpreter" }
 puffin-resolver = { path = "../puffin-resolver" }
@@ -25,6 +26,8 @@ puffin-traits = { path = "../puffin-traits" }
 pypi-types = { path = "../pypi-types" }
 
 anyhow = { workspace = true }
+fs-err = { workspace = true }
 itertools = { workspace = true }
 tempfile = { workspace = true }
+tokio = { workspace = true }
 tracing = { workspace = true }
