+++ crates/puffin-distribution/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "puffin-installer"
+name = "puffin-distribution"
 version = "0.0.1"
 edition = { workspace = true }
 rust-version = { workspace = true }
@@ -11,25 +11,21 @@ license = { workspace = true }
 
 [dependencies]
 distribution-filename = { path = "../distribution-filename" }
-install-wheel-rs = { path = "../install-wheel-rs", default-features = false }
-pep440_rs = { path = "../pep440-rs" }
-pep508_rs = { path = "../pep508-rs" }
-puffin-cache = { path = "../puffin-cache" }
-puffin-client = { path = "../puffin-client" }
 distribution-types = { path = "../distribution-types" }
-puffin-distribution = { path = "../puffin-distribution" }
+install-wheel-rs = { path = "../install-wheel-rs" }
+platform-tags = { path = "../platform-tags" }
+puffin-client = { path = "../puffin-client" }
 puffin-git = { path = "../puffin-git" }
-puffin-interpreter = { path = "../puffin-interpreter" }
-puffin-normalize = { path = "../puffin-normalize" }
 puffin-traits = { path = "../puffin-traits" }
 pypi-types = { path = "../pypi-types" }
 
 anyhow = { workspace = true }
+bytesize = { workspace = true }
 fs-err = { workspace = true }
-fxhash = { workspace = true }
 rayon = { workspace = true }
 tempfile = { workspace = true }
-thiserror = { workspace = true }
 tokio = { workspace = true }
+tokio-util = { workspace = true, features = ["compat"] }
 tracing = { workspace = true }
 url = { workspace = true }
+zip = { workspace = true }
