+++ crates/puffin-installer/src/downloader.rs
@@ -3,28 +3,19 @@ use std::path::{Path, PathBuf};
 use std::sync::Arc;
 
 use anyhow::{bail, Result};
-use bytesize::ByteSize;
-use tempfile::TempDir;
 use tokio::task::JoinSet;
-use tokio_util::compat::FuturesAsyncReadCompatExt;
-use tracing::debug;
-use url::Url;
 
-use distribution_types::direct_url::{DirectArchiveUrl, DirectGitUrl};
-use distribution_types::{BuiltDist, Dist, RemoteSource, SourceDist};
+use distribution_types::{Dist, RemoteSource};
 use puffin_client::RegistryClient;
-use puffin_git::GitSource;
+use puffin_distribution::{Download, Fetcher};
 
 use crate::locks::Locks;
-
-const GIT_CACHE: &str = "git-v0";
 
 pub struct Downloader<'a> {
     client: &'a RegistryClient,
     cache: &'a Path,
     locks: Arc<Locks>,
     reporter: Option<Box<dyn Reporter>>,
-    /// Block building source distributions by not downloading them
     no_build: bool,
 }
 
@@ -49,7 +40,7 @@ impl<'a> Downloader<'a> {
         }
     }
 
-    /// Optionally, block downloading source distributions
+    /// Optionally, block downloading source distributions.
     #[must_use]
     pub fn with_no_build(self, no_build: bool) -> Self {
         Self { no_build, ..self }
@@ -109,144 +100,8 @@ async fn fetch(
 ) -> Result<Download> {
     let lock = locks.acquire(&dist).await;
     let _guard = lock.lock().await;
-
-    match &dist {
-        Dist::Built(BuiltDist::Registry(wheel)) => {
-            // Fetch the wheel.
-            let url = Url::parse(&wheel.file.url)?;
-            let reader = client.stream_external(&url).await?;
-
-            // If the file is greater than 5MB, write it to disk; otherwise, keep it in memory.
-            let small_size = if let Some(size) = wheel.file.size {
-                let byte_size = ByteSize::b(size as u64);
-                if byte_size < ByteSize::mb(5) {
-                    Some(size)
-                } else {
-                    None
-                }
-            } else {
-                None
-            };
-            if let Some(small_size) = small_size {
-                debug!(
-                    "Fetching in-memory wheel from registry: {dist} ({})",
-                    ByteSize::b(small_size as u64)
-                );
-
-                // Read into a buffer.
-                let mut buffer = Vec::with_capacity(small_size);
-                let mut reader = tokio::io::BufReader::new(reader.compat());
-                tokio::io::copy(&mut reader, &mut buffer).await?;
-
-                Ok(Download::Wheel(WheelDownload::InMemory(InMemoryWheel {
-                    dist,
-                    buffer,
-                })))
-            } else {
-                let size = small_size.map_or("unknown size".to_string(), |size| size.to_string());
-                debug!("Fetching disk-based wheel from registry: {dist} ({size})");
-
-                // Download the wheel to a temporary file.
-                let temp_dir = tempfile::tempdir_in(cache)?;
-                let wheel_filename = &wheel.file.filename;
-                let wheel_file = temp_dir.path().join(wheel_filename);
-                let mut writer = tokio::fs::File::create(&wheel_file).await?;
-                tokio::io::copy(&mut reader.compat(), &mut writer).await?;
-
-                Ok(Download::Wheel(WheelDownload::Disk(DiskWheel {
-                    dist,
-                    path: wheel_file,
-                    temp_dir: Some(temp_dir),
-                })))
-            }
-        }
-
-        Dist::Built(BuiltDist::DirectUrl(wheel)) => {
-            debug!("Fetching disk-based wheel from URL: {}", &wheel.url);
-
-            // Fetch the wheel.
-            let reader = client.stream_external(&wheel.url).await?;
-
-            // Download the wheel to a temporary file.
-            let temp_dir = tempfile::tempdir_in(cache)?;
-            let wheel_filename = wheel.filename()?;
-            let wheel_file = temp_dir.path().join(wheel_filename);
-            let mut writer = tokio::fs::File::create(&wheel_file).await?;
-            tokio::io::copy(&mut reader.compat(), &mut writer).await?;
-
-            Ok(Download::Wheel(WheelDownload::Disk(DiskWheel {
-                dist,
-                path: wheel_file,
-                temp_dir: Some(temp_dir),
-            })))
-        }
-
-        Dist::Source(SourceDist::Registry(sdist)) => {
-            debug!(
-                "Fetching source distribution from registry: {}",
-                &sdist.file.url
-            );
-
-            let url = Url::parse(&sdist.file.url)?;
-            let reader = client.stream_external(&url).await?;
-
-            // Download the source distribution.
-            let temp_dir = tempfile::tempdir_in(cache)?;
-            let sdist_filename = sdist.filename()?;
-            let sdist_file = temp_dir.path().join(sdist_filename);
-            let mut writer = tokio::fs::File::create(&sdist_file).await?;
-            tokio::io::copy(&mut reader.compat(), &mut writer).await?;
-
-            Ok(Download::SourceDist(SourceDistDownload {
-                dist,
-                sdist_file,
-                subdirectory: None,
-                temp_dir: Some(temp_dir),
-            }))
-        }
-
-        Dist::Source(SourceDist::DirectUrl(sdist)) => {
-            debug!("Fetching source distribution from URL: {}", sdist.url);
-
-            let DirectArchiveUrl { url, subdirectory } = DirectArchiveUrl::from(&sdist.url);
-
-            let reader = client.stream_external(&url).await?;
-            let mut reader = tokio::io::BufReader::new(reader.compat());
-
-            // Download the source distribution.
-            let temp_dir = tempfile::tempdir_in(cache)?;
-            let sdist_filename = sdist.filename()?;
-            let sdist_file = temp_dir.path().join(sdist_filename);
-            let mut writer = tokio::fs::File::create(&sdist_file).await?;
-            tokio::io::copy(&mut reader, &mut writer).await?;
-
-            Ok(Download::SourceDist(SourceDistDownload {
-                dist,
-                sdist_file,
-                subdirectory,
-                temp_dir: Some(temp_dir),
-            }))
-        }
-
-        Dist::Source(SourceDist::Git(sdist)) => {
-            debug!("Fetching source distribution from Git: {}", sdist.url);
-
-            let DirectGitUrl { url, subdirectory } = DirectGitUrl::try_from(&sdist.url)?;
-
-            let git_dir = cache.join(GIT_CACHE);
-            let source = GitSource::new(url, git_dir);
-            let sdist_file = tokio::task::spawn_blocking(move || source.fetch())
-                .await??
-                .into();
-
-            Ok(Download::SourceDist(SourceDistDownload {
-                dist,
-                sdist_file,
-                subdirectory,
-                temp_dir: None,
-            }))
-        }
-    }
+    let metadata = Fetcher::new(&cache).fetch_dist(&dist, &client).await?;
+    Ok(metadata)
 }
 
 pub trait Reporter: Send + Sync {
@@ -255,88 +110,4 @@ pub trait Reporter: Send + Sync {
 
     /// Callback to invoke when the operation is complete.
     fn on_download_complete(&self);
-}
-
-/// A downloaded wheel that's stored in-memory.
-#[derive(Debug)]
-pub struct InMemoryWheel {
-    /// The remote distribution from which this wheel was downloaded.
-    pub(crate) dist: Dist,
-    /// The contents of the wheel.
-    pub(crate) buffer: Vec<u8>,
-}
-
-/// A downloaded wheel that's stored on-disk.
-#[derive(Debug)]
-pub struct DiskWheel {
-    /// The remote distribution from which this wheel was downloaded.
-    pub(crate) dist: Dist,
-    /// The path to the downloaded wheel.
-    pub(crate) path: PathBuf,
-    /// The download location, to be dropped after use.
-    #[allow(dead_code)] // We only want the drop implementation
-    pub(crate) temp_dir: Option<TempDir>,
-}
-
-/// A downloaded wheel.
-#[derive(Debug)]
-pub enum WheelDownload {
-    InMemory(InMemoryWheel),
-    Disk(DiskWheel),
-}
-
-impl WheelDownload {
-    /// Return the [`Dist`] from which this wheel was downloaded.
-    pub fn remote(&self) -> &Dist {
-        match self {
-            WheelDownload::InMemory(wheel) => &wheel.dist,
-            WheelDownload::Disk(wheel) => &wheel.dist,
-        }
-    }
-}
-
-/// A downloaded source distribution.
-#[derive(Debug)]
-pub struct SourceDistDownload {
-    /// The remote distribution from which this source distribution was downloaded.
-    pub(crate) dist: Dist,
-    /// The path to the downloaded archive or directory.
-    pub(crate) sdist_file: PathBuf,
-    /// The subdirectory within the archive or directory.
-    pub(crate) subdirectory: Option<PathBuf>,
-    /// We can't use source dist archives, we build them into wheels which we persist and then drop
-    /// the source distribution. This field is non for git dependencies, which we keep in the cache.
-    #[allow(dead_code)] // We only keep it for the drop impl
-    pub(crate) temp_dir: Option<TempDir>,
-}
-
-/// A downloaded distribution, either a wheel or a source distribution.
-#[derive(Debug)]
-pub enum Download {
-    Wheel(WheelDownload),
-    SourceDist(SourceDistDownload),
-}
-
-impl std::fmt::Display for Download {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Download::Wheel(wheel) => write!(f, "{wheel}"),
-            Download::SourceDist(sdist) => write!(f, "{sdist}"),
-        }
-    }
-}
-
-impl std::fmt::Display for WheelDownload {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            WheelDownload::InMemory(wheel) => write!(f, "{}", wheel.dist),
-            WheelDownload::Disk(wheel) => write!(f, "{}", wheel.dist),
-        }
-    }
-}
-
-impl std::fmt::Display for SourceDistDownload {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.dist)
-    }
 }
