+++ crates/puffin-installer/src/unzipper.rs
@@ -1,18 +1,13 @@
 use std::cmp::Reverse;
-use std::io::{Read, Seek};
 use std::path::Path;
 
 use anyhow::Result;
-use rayon::iter::ParallelBridge;
-use rayon::iter::ParallelIterator;
 use tracing::debug;
-use zip::ZipArchive;
 
-use distribution_types::{CachedDist, Dist, Identifier};
+use distribution_types::{CachedDist, Dist, Identifier, RemoteSource};
+use puffin_distribution::{Unzip, WheelDownload};
 
 use crate::cache::WheelCache;
-use crate::downloader::WheelDownload;
-use crate::vendor::{CloneableSeekableReader, HasLength};
 
 #[derive(Default)]
 pub struct Unzipper {
@@ -40,10 +35,8 @@ impl Unzipper {
 
         // Sort the wheels by size.
         let mut downloads = downloads;
-        downloads.sort_unstable_by_key(|wheel| match wheel {
-            WheelDownload::Disk(_) => Reverse(usize::MIN),
-            WheelDownload::InMemory(wheel) => Reverse(wheel.buffer.len()),
-        });
+        downloads
+            .sort_unstable_by_key(|wheel| Reverse(wheel.remote().size().unwrap_or(usize::MIN)));
 
         let staging = tempfile::tempdir_in(wheel_cache.root())?;
 
@@ -57,7 +50,7 @@ impl Unzipper {
             // Unzip the wheel.
             tokio::task::spawn_blocking({
                 let target = staging.path().join(remote.distribution_id());
-                move || unzip_wheel(download, &target)
+                move || download.unzip(&target)
             })
             .await??;
 
@@ -91,60 +84,6 @@ impl Unzipper {
 
         Ok(wheels)
     }
-}
-
-/// Unzip a wheel into the target directory.
-fn unzip_wheel(wheel: WheelDownload, target: &Path) -> Result<()> {
-    match wheel {
-        WheelDownload::InMemory(wheel) => unzip_archive(std::io::Cursor::new(wheel.buffer), target),
-        WheelDownload::Disk(wheel) => unzip_archive(fs_err::File::open(wheel.path)?, target),
-    }
-}
-
-/// Unzip a zip archive into the target directory.
-fn unzip_archive<R: Send + Read + Seek + HasLength>(reader: R, target: &Path) -> Result<()> {
-    // Unzip in parallel.
-    let archive = ZipArchive::new(CloneableSeekableReader::new(reader))?;
-    (0..archive.len())
-        .par_bridge()
-        .map(|file_number| {
-            let mut archive = archive.clone();
-            let mut file = archive.by_index(file_number)?;
-
-            // Determine the path of the file within the wheel.
-            let file_path = match file.enclosed_name() {
-                Some(path) => path.to_owned(),
-                None => return Ok(()),
-            };
-
-            // Create necessary parent directories.
-            let path = target.join(file_path);
-            if file.is_dir() {
-                fs_err::create_dir_all(path)?;
-                return Ok(());
-            }
-            if let Some(parent) = path.parent() {
-                fs_err::create_dir_all(parent)?;
-            }
-
-            // Write the file.
-            let mut outfile = fs_err::File::create(&path)?;
-            std::io::copy(&mut file, &mut outfile)?;
-
-            // Set permissions.
-            #[cfg(unix)]
-            {
-                use std::fs::Permissions;
-                use std::os::unix::fs::PermissionsExt;
-
-                if let Some(mode) = file.unix_mode() {
-                    std::fs::set_permissions(&path, Permissions::from_mode(mode))?;
-                }
-            }
-
-            Ok(())
-        })
-        .collect::<Result<_>>()
 }
 
 pub trait Reporter: Send + Sync {
