+++ crates/puffin-macros/Cargo.toml
@@ -10,7 +10,5 @@ authors = { workspace = true }
 license = { workspace = true }
 
 [dependencies]
-colored = { workspace = true }
 fxhash = { workspace = true }
 once_cell = { workspace = true }
-tracing = { workspace = true }
