+++ crates/puffin-resolver/Cargo.toml
@@ -19,6 +19,7 @@ platform-tags = { path = "../platform-tags" }
 puffin-cache = { path = "../puffin-cache" }
 puffin-client = { path = "../puffin-client" }
 distribution-types = { path = "../distribution-types" }
+puffin-distribution = { path = "../puffin-distribution" }
 puffin-git = { path = "../puffin-git" }
 puffin-interpreter = { path = "../puffin-interpreter" }
 puffin-macros = { path = "../puffin-macros" }
