--- crates/puffin-resolver/src/distribution/mod.rs
@@ -1,6 +1,0 @@
-pub(crate) use built_dist::BuiltDistFetcher;
-pub(crate) use source_dist::{Reporter as SourceDistributionReporter, SourceDistFetcher};
-
-mod built_dist;
-mod cached_wheel;
-mod source_dist;
